Adding auto detect
Will now use a little exec'ing to determine your package maanger. Based
on that we will know what/how to parse the contents coming in.

Updated flags as well to deprecate `-os` flag since really there is no
need for it to live much longer with `-autodetect` now being possible.