Fix issue with new autodetect being default
Previously we wanted to keep the program going if auto detect could not
find the right program since you might have passed it in with the --os
flag.

Now that the behavior has changed if auto detect couldn't actually work
we just want to stop it and not let it continue.

The issue in the autodetect docker file was that which was not
installed. So I updated the docs/dockerfile to actually have the
requirements needed.